Move `BannerManager` to `com.osudroid` package
